+++ src/majordomo_llm/providers/cohere.py
@@ -3,10 +3,8 @@
 import json
 import os
 import time
-from typing import Type
 
 import cohere
-from pydantic import BaseModel
 from tenacity import (
     retry,
     stop_after_attempt,
@@ -166,7 +164,7 @@ class Cohere(LLM):
 
     async def _get_structured_response(
         self,
-        response_model: Type[T],
+        response_model: type[T],
         user_prompt: str,
         system_prompt: str | None = None,
         temperature: float = 0.3,